Fix persistant data for urlsToConsult
