+++ pix-editor/app/templates/authenticated/competence/prototypes/single.hbs
@@ -64,6 +64,9 @@
                      @setDisplaySolutionToDisplayField={{this.setDisplaySolutionToDisplayField}}
                      @removeIllustration={{this.removeIllustration}}
                      @removeAttachment={{this.removeAttachment}}
+                    @invalidUrlsToConsult={{this.invalidUrlsToConsult}}
+                    @urlsToConsult={{this.urlsToConsult}}
+                    @setUrlsToConsult={{this.setUrlsToConsult}}
     />
   </div>
 <div class="ui vertical compact labeled icon menu challenge-menu">
