+++ pix-editor/tests/integration/components/form/challenge-test.js
@@ -83,63 +83,4 @@ module('Integration | Component | challenge-form', function(hooks) {
     await new Promise(resolve => setTimeout(resolve, 400));
     await settled();
   });
-
-  module('when `edition` is `false`', function () {
-    test('it should display a list of url to consult when challenge have', async function(assert) {
-      // given
-      const store = this.owner.lookup('service:store');
-      const challengeData = store.createRecord('challenge',{
-        id: 'recChallenge0',
-        genealogy: 'Prototype 1',
-        urlsToConsult: ['jean', 'jacques', 'azimut']
-      });
-      this.set('challengeData', challengeData);
-
-      // When
-      await render(hbs`<Form::Challenge @challenge={{this.challengeData}} @edition={{false}}/>`);
-
-      // then
-      const urlsToConsultInput = await find('[data-test-challenge-urls-to-consult] input');
-
-      assert.strictEqual(urlsToConsultInput.value, 'jean, jacques, azimut');
-    });
-
-    test('it should not display a list of url to consult when challenge have not', async function(assert) {
-      const store = this.owner.lookup('service:store');
-      const challengeData = store.createRecord('challenge',{
-        id: 'recChallenge0',
-        genealogy: 'Prototype 1',
-        urlsToConsult: null
-      });
-      this.set('challengeData', challengeData);
-
-      // When
-      await render(hbs`<Form::Challenge @challenge={{this.challengeData}} @edition={{false}}/>`);
-
-      // then
-      assert.dom('[data-test-challenge-urls-to-consult]').doesNotExist();
-    });
-  });
-
-  module('when `edition` is `true`', function() {
-    test('it should filter invalid url and display it', async function(assert) {
-      // given
-
-      const store = this.owner.lookup('service:store');
-      const challengeData = store.createRecord('challenge',{
-        id: 'recChallenge0',
-        genealogy: 'Prototype 1',
-        urlsToConsult: ['https://toot.fr']
-      });
-      this.set('challengeData', challengeData);
-
-      // When
-      await render(hbs`<Form::Challenge @challenge={{this.challengeData}} @edition={{true}}/>`);
-      await fillByLabel('URLs externes à consulter :', 'https://toot.fr, toot.fr, http://toot.fr');
-
-      // then
-      assert.dom('[data-test-invalid-urls-to-consult]').hasText('URLs invalides : toot.fr');
-      assert.deepEqual(challengeData.urlsToConsult, ['https://toot.fr', 'http://toot.fr']);
-    });
-  });
 });
